requirements.txt
Django>=5.0,<6.0
python-dotenv
djangorestframework>=3.14
Pillow
celery
redis
python-decouple
python-jose[cryptography]
django-cors-headers
psycopg2-binary
gunicorn
